feat: 폴더 및 경로 추천 기능 기본 구조 추가

+++ src/main/java/Purple/Purple/repository/PlaceRepository.java
@@ -1,7 +1,7 @@
 package Purple.Purple.repository;
 
-import Purple.Purple.domian.Place;
+import Purple.Purple.domain.Place;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlaceRepository extends JpaRepository<Place, Long> {
+public interface PlaceRepository extends JpaRepository<Place, Integer> {
 }
